chatbot_model/gpt_model/requirements.txt
fire>=0.1.3
regex==2017.4.5
requests==2.21.0
tqdm==4.31.1
toposort==1.5
numpy==1.17.3
Pillow==6.2.1
flask
tensorflow-gpu==1.15.0